requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.0
schedule>=1.2.0
